requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pydub==0.25.1
cloudinary==1.41.0 